+++ hudi-demo-common/src/main/java/com/leesf/data/CustomDataGenerator.java
@@ -1,4 +1,4 @@
-package com.leesf.multiversion.utils;
+package com.leesf.data;
 
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.sql.Dataset;
