+++ hudi-demo-common/src/main/java/com/leesf/data/OpType.java
@@ -1,4 +1,4 @@
-package com.leesf.multiversion.utils;
+package com.leesf.data;
 
 public enum OpType {
     UPDATE, INSERT, APPEND, DELETE
