+++ hudi-multiversion-demo/pom.xml
@@ -20,6 +20,12 @@
 
     <dependencies>
         <dependency>
+            <groupId>com.leesf</groupId>
+            <artifactId>hudi-demo-common</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-core_${scala.binary.version}</artifactId>
             <version>${spark.version}</version>
