--- hudi-multiversion-demo/src/main/java/com/leesf/multiversion/MergeOnReadMultiVersionDemo.java
@@ -1,5 +1,0 @@
-package com.leesf.multiversion;
-
-public class MergeOnReadMultiVersionDemo {
-
-}
