--- hudi-partition-demo/src/main/java/com/leesf/utils/OpType.java
@@ -1,5 +1,0 @@
-package com.leesf.utils;
-
-public enum OpType {
-    UPDATE, INSERT, APPEND, DELETE
-}
